Added class Effect choice (pick). Identify Level Stats logic (scriptable object) LevelStats.cs Optimized scenes draw calls with Framedebugger. Cleaned up the project files.

+++ Assets/Scripts/Game/UI/DefenderHud.cs
@@ -42,7 +42,7 @@ using UnityEngine.UI;
         _currentWave = currentWave;
     }
 
-    public void PrepareTime(int prepareTime)
+    public void PrepareTime(float prepareTime)
     {
         if (prepareTime > 0)
         {
